requirements.txt
fastapi
uvicorn
pandas
joblib
scikit-learn
groq
pydantic
python-multipart
python-dotenv
